+++ tests/helpers/test_file_helper.py
@@ -12,3 +12,4 @@ def test_files_in_dir():
     assert join(this_dir, 'test_rich_text_helper.py') in this_dir_files_strs
     assert len(files_in_dir(this_dir, with_extname='illmatic')) == 0
     assert len(files_in_dir(this_dir, with_extname='py')) >= 3
+    assert len(files_in_dir(this_dir, with_extname='.py')) >= 3